adjust replace json patch to replace elements in folders

+++ metadata_backend/api/handlers.py
@@ -1,5 +1,6 @@
 """Handle HTTP methods for server."""
 import json
+import re
 import mimetypes
 from collections import Counter
 from math import ceil
@@ -366,7 +367,7 @@ class RESTApiHandler:
             current_user = req.app["Session"]["user_info"]
             check_user = await user_op.check_user_has_doc(collection, current_user, accession_id)
             if check_user:
-                collection = "drafts"
+                await user_op.remove_objects(current_user, "drafts", [accession_id])
             else:
                 reason = "This object does not seem to belong to anything."
                 LOG.error(reason)
@@ -547,19 +548,24 @@ class RESTApiHandler:
         patch_ops = await self._get_data(req)
         allowed_paths = ["/name", "/description"]
         array_paths = ["/metadataObjects/-", "/drafts/-"]
+        tags_path = re.compile("^/(metadataObjects|drafts)/[0-9]*/(tags)$")
         for op in patch_ops:
-            if all(i not in op["path"] for i in allowed_paths + array_paths):
-                reason = f"Request contains '{op['path']}' key that cannot be updated to folders."
-                LOG.error(reason)
-                raise web.HTTPBadRequest(reason=reason)
-            if op["op"] in ["remove", "copy", "test", "move"]:
-                reason = f"{op['op']} on {op['path']} is not allowed."
-                LOG.error(reason)
-                raise web.HTTPUnauthorized(reason=reason)
-            if op["op"] == "replace" and op["path"] in array_paths:
-                reason = f"{op['op']} on {op['path']} is not allowed."
-                LOG.error(reason)
-                raise web.HTTPUnauthorized(reason=reason)
+            if tags_path.match(op["path"]):
+                LOG.info(f"{op['op']} on tags in folder")
+                pass
+            else:
+                if all(i not in op["path"] for i in allowed_paths + array_paths):
+                    reason = f"Request contains '{op['path']}' key that cannot be updated to folders."
+                    LOG.error(reason)
+                    raise web.HTTPBadRequest(reason=reason)
+                if op["op"] in ["remove", "copy", "test", "move"]:
+                    reason = f"{op['op']} on {op['path']} is not allowed."
+                    LOG.error(reason)
+                    raise web.HTTPUnauthorized(reason=reason)
+                if op["op"] == "replace" and op["path"] in array_paths:
+                    reason = f"{op['op']} on {op['path']}; replacing all objects is not allowed."
+                    LOG.error(reason)
+                    raise web.HTTPUnauthorized(reason=reason)
 
         check_user = await self._handle_check_ownedby_user(req, "folders", folder_id)
         if not check_user:
@@ -722,7 +728,7 @@ class RESTApiHandler:
 
         for folder_id in user["folders"]:
             _folder = await fold_ops.read_folder(folder_id)
-            if not _folder["published"]:
+            if "published" in _folder and not _folder["published"]:
                 for obj in _folder["drafts"] + _folder["metadataObjects"]:
                     await obj_ops.delete_metadata_object(obj["schema"], obj["accessionId"])
                 await fold_ops.delete_folder(folder_id)
